requirements.txt
amulet-core==1.4.6
amulet-nbt==1.0.3.10
numpy==1.21.0
PyMCTranslate==1.0.7
